Separate within different blueprints, score_bets, binary_bets and bets calls
